Change all components to be `index.jsx`-files
The reason for this change is to easily add `index.less` files for each
component which creates a consistent namespace for all files. In
addition this also opens up for `index.test.js` files.

PS! Ran into a weird bug where the previously called 'Sider', would
break the build if it was put inside a folder called 'Sider'. Changed
this folder name to 'Asider', and now it works. Strange.. :weary: