Move to a new styling practice
After not being happy with the stock semantic-ui css styles, I decided to swap them out with a theme called yeti. This is more modern, and hence fits better for the vibe I'm going for.

Decided to move style out of the `.jsx` files and into `.sass` files. This is to organize and use postcss-features like autoprefixer. This commit only moves the styles for `Template`.